refac: reuse `AccountWithSlots` struct for temp storage and cache
